Ajout connexion

+++ app/src/main/java/com/example/tranquitaskapp/User.kt
@@ -3,7 +3,7 @@ package com.example.tranquitaskapp
 object User {
     var mail : String = ""
     var username : String = ""
-    var coins : Int = 0
+    var coins : Long = 0
     var profile_picture = ""
 
     fun getUser():User{
